Segmented, bitstrings, whois, types, logging...
- Fixed whois encoding to write APDU with metadata rather than just two fields from the APDU
- Fixed segmented ComplexAck handling
- Fixed date decoding
- Added bitstring support (encoding, decoding, and transfer to/from struct types)
- Added encoding of octetstring, date, time, bitstring
- Added Decoder.IgnorePropertyAccessError to allow AccessError properties to be ignored rather than blowing up decode
- Added tagNumberPeek and used it to fix empty property values (tag open, tag close, with nothing in between)
- PropertyID is now a proper type - with matching const values for known types
- PDUTypes are now the correct values and are << or >> 4 when encoded/decoded
- Added AnyDayOfWeek const
- Added known ObjectType consts

+++ encoding/readmultiple.go
@@ -45,7 +45,7 @@ func (e *Encoder) properties(properties []bactype.Property) error {
 	for _, prop := range properties {
 		// Tag 0 - Property ID
 		tag = 0
-		e.contextEnumerated(tag, prop.Type)
+		e.contextEnumerated(tag, uint32(prop.Type))
 
 		// Tag 1 (OPTIONAL) - Array Length
 		if prop.ArrayIndex != ArrayAll {
